src/modules/users/services/UpdateUserAvatarService.spec.ts
import AppError from '@shared/errors/appError';

import FakeStorageProvider from '@shared/container/providers/StorageProviders/fakes/FakeStorageProvider';
import FakeUsersRepository from '../repositories/fakes/FakeUsersRepository';
import UpdateUserAvatarService from './UpdateUserAvatarService';

let fakeUsersRepository: FakeUsersRepository;
let fakeStorageProvider: FakeStorageProvider;
let updateUser: UpdateUserAvatarService;

describe('UpdateUser', () => {
  beforeEach(() => {
    fakeUsersRepository = new FakeUsersRepository();
    fakeStorageProvider = new FakeStorageProvider();

    updateUser = new UpdateUserAvatarService(
      fakeUsersRepository,
      fakeStorageProvider,
    );
  });
  it('should be able to update user', async () => {
    const user = await fakeUsersRepository.create({
      name: 'teste',
      email: 'email@teste.com',
      password: '123456',
    });

    await updateUser.execute({
      user_id: user.id,
      fileName: 'avatar.jpg',
    });

    expect(user.avatar).toBe('avatar.jpg');
  });

  it('should not be able to update avatar from non existing user', async () => {
    await expect(
      updateUser.execute({
        user_id: 'non-existing-user',
        fileName: 'avatar.jpg',
      }),
    ).rejects.toBeInstanceOf(AppError);
  });

  it('should delete old avatar when updating new one', async () => {
    const deleteFile = jest.spyOn(fakeStorageProvider, 'deleteFile');

    const user = await fakeUsersRepository.create({
      name: 'teste',
      email: 'email@teste.com',
      password: '123456',
    });

    await updateUser.execute({
      user_id: user.id,
      fileName: 'avatar.jpg',
    });

    await updateUser.execute({
      user_id: user.id,
      fileName: 'avatar2.jpg',
    });

    expect(deleteFile).toHaveBeenCalledWith('avatar.jpg');
    expect(user.avatar).toBe('avatar2.jpg');
  });
});
